grouping worksheets proto
https://app.asana.com/0/1152783500105769/1203549842177806/f

+++ lib/providers/worksheets_provider.dart
@@ -2,6 +2,7 @@
 
 import 'dart:async';
 
+import 'package:collection/collection.dart';
 import 'package:flutter/services.dart' show rootBundle;
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:path/path.dart';
@@ -19,6 +20,7 @@ final worksheetNotifierProvider =
     AsyncNotifierProvider.autoDispose<WorksheetNotifier, List<Worksheet>>(WorksheetNotifier.new);
 
 final worksheetTypeProvider = Provider<WorksheetTypeRepository>((ref) => WorksheetTypeRepository());
+final worksheetRepoProvider = Provider<WorksheetRepository>((ref) => WorksheetRepository(ref));
 
 class WorksheetEventNotifier extends StateNotifier<WorksheetEvent> {
   WorksheetEventNotifier() : super(WorksheetEvent(WorksheetEventType.Default, List.empty()));
@@ -27,10 +29,19 @@ class WorksheetEventNotifier extends StateNotifier<WorksheetEvent> {
 final worksheetEventProvider =
     StateNotifierProvider<WorksheetEventNotifier, WorksheetEvent>((ref) => WorksheetEventNotifier());
 
+final childWorksheetsProvider = FutureProvider.autoDispose.family<List<Worksheet>, int>((ref, id) async {
+  final repo = ref.watch(worksheetRepoProvider);
+  return repo.getChildWorksheets(id);
+});
+
 class WorksheetDb {
   final databaseName = "notes.db";
   final tableName = "notes";
-  static const migrationScripts = ['alter table notes add column is_complete integer default 0;'];
+  static const migrationScripts = [
+    'alter table notes add column is_complete integer default 0;',
+    "alter table notes add column parent_id integer default -1;",
+    "create index idx_notes_parent_id on notes (parent_id);"
+  ];
 
   final fieldMap = {
     "id": "INTEGER PRIMARY KEY AUTOINCREMENT",
@@ -40,7 +51,8 @@ class WorksheetDb {
     "date_last_edited": "INTEGER",
     "note_color": "INTEGER",
     "is_archived": "INTEGER",
-    "is_complete": "INTEGER"
+    "is_complete": "INTEGER",
+    "parent_id": "INTEGER DEFAULT -1"
   };
 
   Database? _database;
@@ -56,7 +68,7 @@ class WorksheetDb {
     var path = await getDatabasesPath();
     var dbPath = join(path, 'notes.db');
     // ignore: argument_type_not_assignable
-    Database dbConnection = await openDatabase(dbPath, version: 2, onCreate: (Database db, int version) async {
+    Database dbConnection = await openDatabase(dbPath, version: 4, onCreate: (Database db, int version) async {
       print("executing create query from onCreate callback");
       await db.execute(_buildCreateQuery());
     }, onUpgrade: (Database db, int oldVersion, int newVersion) async {
@@ -117,32 +129,52 @@ class WorksheetRepository {
     return db.delete("notes", where: "id = ?", whereArgs: [id]);
   }
 
+  Future<Worksheet?> getWorksheet(int id) async {
+    final db = await ref.read(worksheetDbProvider.future);
+    final res = await db.query("notes", where: "id = ?", whereArgs: [id], distinct: true);
+    return res.isNotEmpty ? res.map((worksheet) => Worksheet.fromJson(worksheet)).firstOrNull : null;
+  }
+
   Future<List<Worksheet>> getWorksheets() async {
     final db = await ref.read(worksheetDbProvider.future);
     // query all the worksheets sorted by last edited
-    var res = await db.query("notes", orderBy: "date_last_edited desc", where: "is_archived = ?", whereArgs: [0]);
-    return res.isNotEmpty ? res.map((note) => Worksheet.fromJson(note)).toList() : [];
+    var res = await db.query("notes", orderBy: "date_created desc", where: "is_archived = ?", whereArgs: [0]);
+    return res.isNotEmpty ? res.map((worksheet) => Worksheet.fromJson(worksheet)).toList() : [];
+  }
+
+  Future<List<Worksheet>> getChildWorksheets(int id) async {
+    final db = await ref.read(worksheetDbProvider.future);
+    final res =
+        await db.query("notes", where: "parent_id = ? or id = ?", whereArgs: [id, id], orderBy: "date_created desc");
+    return res.isNotEmpty ? res.map((worksheet) => Worksheet.fromJson(worksheet)).toList() : [];
   }
 }
 
 enum WorksheetEventType { Default, Reloaded, Added, Modified, Archived, Deleted }
 
 class WorksheetEvent {
-  WorksheetEvent(this.type, this.worksheets, {this.worksheet}) : this.timestamp = DateTime.now();
+  WorksheetEvent(this.type, this.worksheets, {this.worksheet, this.worksheetId}) : this.timestamp = DateTime.now();
   final DateTime timestamp;
   final WorksheetEventType type;
   final Worksheet? worksheet;
   final List<Worksheet> worksheets;
+  final int? worksheetId;
 
   bool operator ==(o) => o is WorksheetEvent && o.type == type && o.timestamp == timestamp;
   @override
   int get hashCode => Object.hash(type, timestamp);
+
+  @override
+  String toString() {
+    return "WorksheetEvent(type: $type, timestamp: $timestamp, worksheet: $worksheet, worksheets: $worksheets, worksheetId: $worksheetId)";
+  }
 }
 
 class WorksheetNotifier extends AutoDisposeAsyncNotifier<List<Worksheet>> {
   @override
   FutureOr<List<Worksheet>> build() {
-    return WorksheetRepository(ref).getWorksheets().then((worksheets) {
+    final repo = ref.watch(worksheetRepoProvider);
+    return repo.getWorksheets().then((worksheets) {
       final provider = ref.read(worksheetEventProvider.notifier);
       provider.state = WorksheetEvent(WorksheetEventType.Reloaded, worksheets);
       return worksheets;
@@ -150,7 +182,7 @@ class WorksheetNotifier extends AutoDisposeAsyncNotifier<List<Worksheet>> {
   }
 
   Future<int> addWorksheet(Worksheet worksheet) async {
-    final repo = WorksheetRepository(ref);
+    final repo = ref.watch(worksheetRepoProvider);
     state = const AsyncLoading();
     worksheet.dateLastEdited = DateTime.now();
     final res = await repo.addWorksheet(worksheet);
@@ -162,8 +194,9 @@ class WorksheetNotifier extends AutoDisposeAsyncNotifier<List<Worksheet>> {
         final provider = ref.read(worksheetEventProvider.notifier);
         provider.state = WorksheetEvent(
             worksheet.id == -1 ? WorksheetEventType.Added : WorksheetEventType.Modified, data,
-            worksheet: cloned);
+            worksheet: cloned, worksheetId: cloned.id);
       });
+
       state = worksheets;
     } else {
       throw new WorksheetDbException("Worksheet couldn't be added!");
@@ -172,31 +205,49 @@ class WorksheetNotifier extends AutoDisposeAsyncNotifier<List<Worksheet>> {
   }
 
   Future<void> deleteWorksheet(int id) async {
-    final repo = WorksheetRepository(ref);
+    final repo = ref.watch(worksheetRepoProvider);
     state = const AsyncLoading();
+    final ws = await repo.getWorksheet(id);
+    if (ws == null) {
+      print("Worksheet $id not found!");
+      return;
+    }
     final res = await repo.deleteWorksheet(id);
+
     if (res == 1) {
-      final worksheets = await AsyncValue.guard(repo.getWorksheets);
-      worksheets.whenData((data) {
+      final asyncWorksheets = await AsyncValue.guard(repo.getWorksheets);
+      asyncWorksheets.whenData((worksheets) async {
+        final orphaned = worksheets.where((ws) => ws.parentId == id);
+        await Future.forEach(orphaned, (element) {
+          ws.parentId = -1;
+          repo.addWorksheet(ws);
+        });
         final provider = ref.read(worksheetEventProvider.notifier);
-        provider.state = WorksheetEvent(WorksheetEventType.Deleted, data);
+        provider.state = WorksheetEvent(WorksheetEventType.Deleted, worksheets, worksheetId: id);
+        if (orphaned.isEmpty) {
+          state = asyncWorksheets;
+        }
       });
-      state = worksheets;
     } else {
       throw new WorksheetDbException("Worksheet couldn't be deleted!");
     }
   }
 
   Future<void> archiveWorksheet(Worksheet worksheet) async {
-    final repo = WorksheetRepository(ref);
+    final repo = ref.watch(worksheetRepoProvider);
     state = const AsyncLoading();
     await repo.archiveWorksheet(worksheet);
     final worksheets = await AsyncValue.guard(repo.getWorksheets);
     worksheets.whenData((data) {
       final provider = ref.read(worksheetEventProvider.notifier);
-      provider.state = WorksheetEvent(WorksheetEventType.Archived, data, worksheet: worksheet);
+      provider.state =
+          WorksheetEvent(WorksheetEventType.Archived, data, worksheet: worksheet, worksheetId: worksheet.id);
     });
     state = worksheets;
+  }
+
+  List<Worksheet>? getCachedChildren(int id) {
+    return state.value?.where((e) => e.parentId == id).toList(growable: false);
   }
 }
 
@@ -215,6 +266,10 @@ class WorksheetTypeRepository {
     var doc = loadYaml(await rootBundle.loadString('assets/question_types.yaml')) as Map;
     _worksheets = Map.unmodifiable(doc.map((k, v) => MapEntry(k.toString(), WorksheetContent.fromYamlMap(k, v))));
 
+    return _worksheets;
+  }
+
+  Map<String, WorksheetContent>? getCachedInquiryTypes() {
     return _worksheets;
   }
 }
